+++ docs/html/search/classes_0.js
@@ -1,7 +1,5 @@
 var searchData=
 [
-  ['bmsdevice',['BmsDevice',['../class_bms_device.html',1,'']]],
-  ['bmsdevicereader',['BmsDeviceReader',['../class_bms_device_reader.html',1,'']]],
-  ['bmsdeviceupdater',['BmsDeviceUpdater',['../class_bms_device_updater.html',1,'']]],
-  ['bmsdevicewriter',['BmsDeviceWriter',['../class_bms_device_writer.html',1,'']]]
+  ['bms',['Bms',['../class_bms.html',1,'']]],
+  ['bmsdevice',['BmsDevice',['../class_bms_device.html',1,'']]]
 ];