+++ docs/html/search/classes_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['sdcardstream',['SdCardStream',['../class_sd_card_stream.html',1,'']]]
+  ['raceuputils',['RaceUpUtils',['../class_race_up_utils.html',1,'']]]
 ];