+++ docs/html/search/classes_2.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['utils',['Utils',['../class_utils.html',1,'']]]
+  ['sdcardstream',['SdCardStream',['../class_sd_card_stream.html',1,'']]]
 ];