+++ docs/html/search/functions_0.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['bmsdeviceread',['bmsDeviceRead',['../class_bms_device_reader.html#af56a2eaeb612587f91b9271ffaf8cda0',1,'BmsDeviceReader']]],
-  ['bmsdevicewrite',['bmsDeviceWrite',['../class_bms_device_writer.html#a0f1433d379e82698c1d48abd8e935e13',1,'BmsDeviceWriter']]]
+  ['bms',['Bms',['../class_bms.html#a5d733fdf4429045471ec8fd56f9ae92e',1,'Bms']]],
+  ['bmsdevice',['BmsDevice',['../class_bms_device.html#a4035559c9370045a409d33cb8941a9f3',1,'BmsDevice']]]
 ];