+++ docs/html/search/functions_4.js
@@ -1,8 +1,9 @@
 var searchData=
 [
-  ['getbmsdevicestatus',['getBmsDeviceStatus',['../class_bms_device_reader.html#a824d04ae77bd18902c1445980fe9dc3a',1,'BmsDeviceReader']]],
-  ['getcellvoltage',['getCellVoltage',['../class_bms_device_reader.html#abcc5341d72ed906225e85b5faa5cf6e8',1,'BmsDeviceReader']]],
-  ['getgpaivbatt',['getGpaiVbatt',['../class_bms_device_reader.html#a1f0651bfef7b2235825320e1e96993a0',1,'BmsDeviceReader']]],
-  ['getserialupdate',['getSerialUpdate',['../class_bms_device_updater.html#a9c5b00e6d5f24df9a7adbb2b772cfafd',1,'BmsDeviceUpdater']]],
-  ['gettemperature',['getTemperature',['../class_bms_device_reader.html#ad3a3a7bb7cae1e3a69f092004cb533ed',1,'BmsDeviceReader']]]
+  ['getarduinoboardmodelname',['getArduinoBoardModelName',['../class_race_up_utils.html#a331fe79dbe61bd7cd84a61b9a05c72e4',1,'RaceUpUtils']]],
+  ['getcellvoltage',['getCellVoltage',['../class_bms_device.html#a0bdebab046f8337914ba88fdd20da1bd',1,'BmsDevice']]],
+  ['getgpaivbatt',['getGpaiVbatt',['../class_bms_device.html#a8b11a6a96bc0634e14bb7847b448bab6',1,'BmsDevice']]],
+  ['getserialupdate',['getSerialUpdate',['../class_bms.html#ab068da9c10b0bd7e865cba83fabaaaef',1,'Bms']]],
+  ['getstatus',['getStatus',['../class_bms_device.html#ad72700d4da6cc7d6ef52e76091ab9d9a',1,'BmsDevice']]],
+  ['gettemperature',['getTemperature',['../class_bms_device.html#ac61154d9b8cc11061dcc7d0243ed955f',1,'BmsDevice']]]
 ];