+++ docs/html/search/functions_7.js
@@ -1,12 +1,13 @@
 var searchData=
 [
   ['sdcardstream',['SdCardStream',['../class_sd_card_stream.html#a0972fdeed91408cfd8b9daab58e1ef2a',1,'SdCardStream']]],
-  ['sendpacktemperatureofbmsdevice',['sendPackTemperatureOfBmsDevice',['../class_bms_device_updater.html#ac2bf4bba3100b48c1f8b42d21c8670f1',1,'BmsDeviceUpdater']]],
-  ['sendpackvoltageofbmsdevice',['sendPackVoltageOfBmsDevice',['../class_bms_device_updater.html#a655ccc4e5f315c383d066c502b73b9e2',1,'BmsDeviceUpdater']]],
-  ['sendsegmentstatus',['sendSegmentStatus',['../class_bms_device_updater.html#a6ddc61da65da16ff6c24995f43f0a6c1',1,'BmsDeviceUpdater']]],
-  ['sendsegmenttemperature',['sendSegmentTemperature',['../class_bms_device_updater.html#aae3a833fa2bd8403d9b55c358a6d2820',1,'BmsDeviceUpdater']]],
-  ['sendsegmentvoltage',['sendSegmentVoltage',['../class_bms_device_updater.html#ab1873577d7413c453222da23b8f6a1ac',1,'BmsDeviceUpdater']]],
-  ['sendstatus',['sendStatus',['../class_bms_device_updater.html#a5cc748a486645d1af6f675274f0d29ae',1,'BmsDeviceUpdater']]],
-  ['sendupdate',['sendUpdate',['../class_bms_device_updater.html#a9449da7019249fb3ce825c12fb61c134',1,'BmsDeviceUpdater']]],
+  ['sendpacktemperatureofbmsdevice',['sendPackTemperatureOfBmsDevice',['../class_bms.html#a49cc5412e466fa94f2e6eb585efb88f7',1,'Bms']]],
+  ['sendpackvoltageofbmsdevice',['sendPackVoltageOfBmsDevice',['../class_bms.html#adf5e55c4858fb8f2b1ecd468f67abed7',1,'Bms']]],
+  ['sendsegmentstatus',['sendSegmentStatus',['../class_bms.html#acb40c808d8fa45d5cf27dabe80c6e508',1,'Bms']]],
+  ['sendsegmenttemperature',['sendSegmentTemperature',['../class_bms.html#a2d3dd5589bbb2db1667510f0b3117e0e',1,'Bms']]],
+  ['sendsegmentvoltage',['sendSegmentVoltage',['../class_bms.html#a1ee6832e0d5662961f15bcb2e9021a44',1,'Bms']]],
+  ['sendstatus',['sendStatus',['../class_bms.html#ac2df5edf6db5cffe6c59967d36f4ab37',1,'Bms']]],
+  ['sendupdate',['sendUpdate',['../class_bms.html#acac85da825d6567cc0b1752c95a48e56',1,'Bms']]],
+  ['setstatus',['setStatus',['../class_bms_device.html#af711c0fbff46642d525c63a668cb7a2d',1,'BmsDevice']]],
   ['setup',['setup',['../class_sd_card_stream.html#ab058227f98edd574579e911170f20464',1,'SdCardStream']]]
 ];